Algorithms

Hadi & Simonoff (1993)

LinRegOutliers.hs93 — Function

hs93(setting; alpha = 0.05, basicsubsetindices = nothing)

Perform the Hadi & Simonoff (1993) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

alpha::Float64: Optional argument of the probability of rejecting the null hypothesis.

basicsubsetindices::Array{Int, 1}: Initial basic subset, by default, the algorithm creates an initial set of clean observations.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> hs93(reg0001)
Dict{Any,Any} with 3 entries:
  "outliers" => [14, 15, 16, 17, 18, 19, 20, 21]
  "t"        => -3.59263
  "d"        => [2.04474, 1.14495, -0.0633255, 0.0632934, -0.354349, -0.766818, -1.06862, -1.47638, -0.7…

References

Hadi, Ali S., and Jeffrey S. Simonoff. "Procedures for the identification of multiple outliers in linear models." Journal of the American Statistical Association 88.424 (1993): 1264-1272.

Kianifard & Swallow (1989)

LinRegOutliers.ks89 — Function

ks89(setting; alpha = 0.05)

Perform the Kianifard & Swallow (1989) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

alpha::Float64: Optional argument of the probability of rejecting the null hypothesis.

Examples

julia> reg0001 = createRegressionSetting(@formula(stackloss ~ airflow + watertemp + acidcond), stackloss)
julia> ks89(reg0001)
2-element Array{Int64,1}:
  4
 21

References

Kianifard, Farid, and William H. Swallow. "Using recursive residuals, calculated on adaptively-ordered observations, to identify outliers in linear regression." Biometrics (1989): 571-585.

Sebert & Montgomery & Rollier (1998)

LinRegOutliers.smr98 — Function

smr98(setting)

Perform the Sebert, Monthomery and Rollier (1998) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> smr98(reg0001)
10-element Array{Int64,1}:
 15
 16
 17
 18
 19
 20
 21
 22
 23
 24

References

Sebert, David M., Douglas C. Montgomery, and Dwayne A. Rollier. "A clustering algorithm for identifying multiple outliers in linear regression." Computational statistics & data analysis 27.4 (1998): 461-484.

Least Median of Squares

LinRegOutliers.lms — Function

lms(setting; iters = nothing, crit = 2.5)

Perform Least Median of Squares regression estimator with random sampling.

Arguments

setting::RegressionSetting: A regression setting object.

iters::Int: Number of random samples.

crit::Float64: Critical value for standardized residuals.

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);

julia> lms(reg)
Dict{Any,Any} with 6 entries:
  "stdres"    => [2.28328, 1.55551, 0.573308, 0.608843, 0.220321, -0.168202, -0.471913, -0.860435, -0.31603, -0.110871  …  85.7265, 88.9849, 103.269, 116.705, 135.229, 159.69,…
  "S"         => 1.17908
  "outliers"  => [14, 15, 16, 17, 18, 19, 20, 21]
  "objective" => 0.515348
  "coef"      => [-56.1972, 1.1581]
  "crit"      => 2.5

References

Rousseeuw, Peter J. "Least median of squares regression." Journal of the American statistical association 79.388 (1984): 871-880.

Least Trimmed Squares

LinRegOutliers.lts — Function

lts(setting; iters, crit)

Perform the Fast-LTS (Least Trimmed Squares) algorithm for a given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

iters::Int: Number of iterations.

crit::Float64: Critical value.

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);
julia> lts(reg)
Dict{Any,Any} with 6 entries:
  "betas"            => [-56.5219, 1.16488]
  "S"                => 1.10918
  "hsubset"          => [11, 10, 5, 6, 23, 12, 13, 9, 24, 7, 3, 4, 8]
  "outliers"         => [14, 15, 16, 17, 18, 19, 20, 21]
  "scaled.residuals" => [2.41447, 1.63472, 0.584504, 0.61617, 0.197052, -0.222066, -0.551027, -0.970146, -0.397538, -0.185558  …  …
  "objective"        => 3.43133

References

Rousseeuw, Peter J., and Katrien Van Driessen. "An algorithm for positive-breakdown regression based on concentration steps." Data Analysis. Springer, Berlin, Heidelberg, 2000. 335-346.

Minimum Volume Ellipsoid (MVE)

LinRegOutliers.mve — Function

mve(data; alpha = 0.01)

Performs the Minimum Volume Ellipsoid algorithm for a robust covariance matrix.

Arguments

data::DataFrame: Multivariate data.

alpha::Float64: Probability for quantiles of Chi-Squared statistic.

References

Van Aelst, Stefan, and Peter Rousseeuw. "Minimum volume ellipsoid." Wiley Interdisciplinary Reviews: Computational Statistics 1.1 (2009): 71-82.

MVE & LTS Plot

LinRegOutliers.mveltsplot — Function

mveltsplot(setting; alpha = 0.05, showplot = true)

Generate MVE - LTS plot for visual detecting of regression outliers.

Arguments

setting::RegressionSetting: A regression setting object.

alpha::Float64: Probability for quantiles of Chi-Squared statistic.

showplot::Bool: Whether a plot is shown or only return statistics.

References

Van Aelst, Stefan, and Peter Rousseeuw. "Minimum volume ellipsoid." Wiley Interdisciplinary Reviews: Computational Statistics 1.1 (2009): 71-82.

Billor & Chatterjee & Hadi (2006)

LinRegOutliers.bch — Function

bch(setting; alpha = 0.05, maxiter = 1000, epsilon = 0.000001)

Perform the Billor & Chatterjee & Hadi (2006) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

alpha::Float64: Optional argument of the probability of rejecting the null hypothesis.

maxiter::Int: Maximum number of iterations for calculating iterative weighted least squares estimates.

epsilon::Float64: Accuracy for determining convergency.

Examples

julia> reg  = createRegressionSetting(@formula(calls ~ year), phones);
julia> Dict{Any,Any} with 7 entries:
"betas"                               => [-55.9205, 1.15572]
"squared.normalized.robust.distances" => [0.104671, 0.0865052, 0.0700692, 0.0553633, 0.0423875, 0.03…
"weights"                             => [0.00186158, 0.00952088, 0.0787321, 0.0787321, 0.0787321, 0…
"outliers"                            => [1, 14, 15, 16, 17, 18, 19, 20, 21]
"squared.normalized.residuals"        => [5.53742e-5, 2.42977e-5, 2.36066e-6, 2.77706e-6, 1.07985e-7…
"residuals"                           => [2.5348, 1.67908, 0.523367, 0.567651, 0.111936, -0.343779, …
"basic.subset"                        => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10  …  15, 16, 17, 18, 19, 20, …

References

Billor, Nedret, Samprit Chatterjee, and Ali S. Hadi. "A re-weighted least squares method for robust regression estimation." American journal of mathematical and management sciences 26.3-4 (2006): 229-252.

Pena & Yohai (1995)

LinRegOutliers.py95 — Function

py95(setting)

Perform the Pena & Yohai (1995) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> reg0001 = createRegressionSetting(@formula(y ~ x1 + x2 + x3), hbk);
julia> py95(reg0001)
ict{Any,Any} with 2 entries:
  "outliers"       => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
  "suspected.sets" => Set([[14, 13], [43, 54, 24, 38, 22], [6, 10], [14, 7, 8, 3, 10, 2, 5, 6, 1, 9, 4…

References

Peña, Daniel, and Victor J. Yohai. "The detection of influential subsets in linear regression by using an influence matrix." Journal of the Royal Statistical Society: Series B (Methodological) 57.1 (1995): 145-156.

## Satman (2013)

LinRegOutliers.satman2013 — Function

satman2013(setting)

Perform Satman (2013) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> eg0001 = createRegressionSetting(@formula(y ~ x1 + x2 + x3), hbk);
julia> satman2013(reg0001)
Dict{Any,Any} with 1 entry:
  "outliers" => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 47]

References

Satman, Mehmet Hakan. "A new algorithm for detecting outliers in linear regression." International Journal of statistics and Probability 2.3 (2013): 101.

## Satman (2015)

LinRegOutliers.satman2015 — Function

satman2013(setting)

Perform Satman (2015) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> eg0001 = createRegressionSetting(@formula(y ~ x1 + x2 + x3), hbk);
julia> satman2013(reg0001)
Dict{Any,Any} with 1 entry:
  "outliers" => [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 14, 47]

References

Satman, Mehmet Hakan. "A new algorithm for detecting outliers in linear regression." International Journal of statistics and Probability 2.3 (2013): 101.

## Setan & Halim & Mohd (2000)

LinRegOutliers.asm2000 — Function

asm2000(setting)

Perform the Setan, Halim and Mohd (2000) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> asm2000(reg0001)
Dict{Any,Any} with 1 entry:
  "outliers" => [15, 16, 17, 18, 19, 20]

References

Setan, Halim, and Mohd Nor Mohamad. "Identifying multiple outliers in linear regression: Robust fit and clustering approach." (2000).

## Least Absolute Deviations (LAD)

LinRegOutliers.lad — Function

lad(setting; starting_betas = nothing)

Perform Least Absolute Deviations regression for a given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

starting_betas::Array{Float64,1}: Starting values of parameter estimations that fed to local search optimizer.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> lad(reg0001)
Dict{Any,Any} with 2 entries:
  "betas"     => [-57.3269, 1.19155]
  "residuals" => [2.14958, 1.25803, 0.0664872, 0.0749413, -0.416605, -0.90815, -1.2997, -1.79124,…

lad(X, y; starting_betas = nothing)

Perform Least Absolute Deviations regression for a given regression setting.

Arguments

X::Array{Float64, 2}: Design matrix of the linear model.

y::Array{Float64, 1}: Response vector of the linear model.

starting_betas::Array{Float64,1}: Starting values of parameter estimations that fed to local search optimizer.

Least Trimmed Absolute Deviations (LTA)

LinRegOutliers.lta — Function

lta(setting; exact = false)

Perform the Hawkins & Olive (1999) algorithm (Least Trimmed Absolute Deviations) for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

exact::Bool: Consider all possible subsets of p or not where p is the number of regression parameters.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> lta(reg0001)
Dict{Any,Any} with 2 entries:
  "betas"     => [-55.5, 1.15]
  "objective" => 5.7

julia> lta(reg0001, exact = true)
Dict{Any,Any} with 2 entries:
  "betas"     => [-55.5, 1.15]
  "objective" => 5.7  

References

Hawkins, Douglas M., and David Olive. "Applications and algorithms for least trimmed sum of absolute deviations regression." Computational Statistics & Data Analysis 32.2 (1999): 119-134.

lta(X, y; exact = false)

Perform the Hawkins & Olive (1999) algorithm (Least Trimmed Absolute Deviations) for the given regression setting.

Arguments

X::Array{Float64, 2}: Design matrix of linear regression model.

y::Array{Float64, 1}: Response vector of linear regression model.

exact::Bool: Consider all possible subsets of p or not where p is the number of regression parameters.

References

Hawkins, Douglas M., and David Olive. "Applications and algorithms for least trimmed sum of absolute deviations regression." Computational Statistics & Data Analysis 32.2 (1999): 119-134.

Hadi (1992)

LinRegOutliers.hadi1992 — Function

hadi1992(multivariateData)

Perform Hadi (1992) algorithm for a given multivariate data.

# Arguments -multivariate::Array{Float64, 2}: Multivariate data.

Examples

julia> multidata = hcat(hbk.x1, hbk.x2, hbk.x3);

julia> hadi1992(multidata)
Dict{Any,Any} with 3 entries:
  "outliers"              => [2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
  "criticial.chi.squared" => 7.81473
  "rth.robust.distance"   => 5.04541

# Reference Hadi, Ali S. "Identifying multiple outliers in multivariate data." Journal of the Royal Statistical Society: Series B (Methodological) 54.3 (1992): 761-771.

Marchette & Solka (2003) Data Images

LinRegOutliers.dataimage — Function

dataimage(dataMatrix; distance = :mahalanobis)

Generate the Marchette & Solka (2003) data image for a given data matrix.

Arguments

dataMatrix::Array{Float64, 1}: Data matrix with dimensions n x p, where n is the number of observations and p is the number of variables.

distance::Symbol: Optional argument for the distance function.

Notes

distance is :mahalanobis by default, for the Mahalanobis distances. 
use 

    dataimage(mat, distance = :euclidean)

to use Euclidean distances.

Examples

julia> x1 = hbk[:,"x1"];
julia> x2 = hbk[:,"x2"];
julia> x3 = hbk[:,"x3"];
julia> mat = hcat(x1, x2, x3);
julia> dataimage(mat)

References

Marchette, David J., and Jeffrey L. Solka. "Using data images for outlier detection." Computational Statistics & Data Analysis 43.4 (2003): 541-552.

Satman's GA based LTS estimation (2012)

LinRegOutliers.galts — Function

galts(setting)

Perform Satman(2012) algorithm for estimating LTS coefficients.

Arguments

setting: A regression setting object.

Examples

Examples

julia> reg = createRegressionSetting(@formula(calls ~ year), phones);
julia> galts(reg)
Dict{Any,Any} with 3 entries:
  "betas"       => [-56.5219, 1.16488]
  "best.subset" => [3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 23, 24]
  "objective"   => 3.43133

References

Satman, M. Hakan. "A genetic algorithm based modification on the lts algorithm for large data sets." Communications in Statistics-Simulation and Computation 41.5 (2012): 644-652.

## Fischler & Bolles (1981) RANSAC Algorithm

LinRegOutliers.ransac — Function

ransac(setting; t, w=0.5, m=0, k=0, d=0, confidence=0.99)

Run the RANSAC (1981) algorithm for the given regression setting

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and a dataset.

t::Float64: The threshold distance of a sample point to the regression hyperplane to determine if it fits the model well.

w::Float64: The probability of a sample point being inlier, default=0.5.

m::Int: The number of points to sample to estimate the model parameter for each iteration. If set to 0, defaults to picking p points which is the minimum required.

k::Int: The number of iterations to run. If set to 0, is calculated according to the formula given in the paper based on outlier probability and the sample set size.

d::Int: The number of close data points required to accept the model. Defaults to number of data points multiplied by inlier ratio.

confidence::Float64: Required to determine the number of optimum iterations if k is not specified.

Examples

julia> df = DataFrame(y=[0,1,2,3,3,4,10], x=[0,1,2,2,3,4,2])
julia> reg = createRegressionSetting(@formula(y ~ x), df)
julia> ransac(reg, t=0.8, w=0.85)
1-element Array{Int64,1}:
 7

References

Martin A. Fischler & Robert C. Bolles (June 1981). "Random Sample Consensus: A Paradigm for Model Fitting with Applications to Image Analysis and Automated Cartography" Comm. ACM. 24 (6): 381–395.

## Minimum Covariance Determinant Estimator (MCD)

LinRegOutliers.mcd — Function

mcd(data; alpha = 0.01)

Performs the Minimum Covariance Determinant algorithm for a robust covariance matrix.

Arguments

data::DataFrame: Multivariate data.

alpha::Float64: Probability for quantiles of Chi-Squared statistic.

# Notes Algorithm is implemented using concentration steps as described in the reference paper. However, details about number of iterations are slightly different.

References

Rousseeuw, Peter J., and Katrien Van Driessen. "A fast algorithm for the minimum covariance determinant estimator." Technometrics 41.3 (1999): 212-223.

## Imon (2005) Algorithm

LinRegOutliers.imon2005 — Function

imon2005(setting)

Perform the Imon 2005 algorithm for a given regression setting.

# Arguments

setting::RegressionSetting: A regression setting.

Notes

The implementation uses LTS rather than LMS as suggested in the paper.

# Reference A. H. M. Rahmatullah Imon (2005) Identifying multiple influential observations in linear regression, Journal of Applied Statistics, 32:9, 929-946, DOI: 10.1080/02664760500163599

Barratt & Angeris & Boyd (2020) CCF algorithm

LinRegOutliers.ccf — Function

ccf(setting; starting_lambdas = nothing)

Perform signed gradient descent for clipped convex functions for a given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

starting_lambdas::Array{Float64,1}: Starting values of weighting parameters used by signed gradient descent.

alpha::Float64: Loss at which a point is labeled as an outlier (points with loss ≥ alpha will be called outliers).

max_iter::Int64: Maximum number of iterations to run signed gradient descent.

beta::Float64: Step size parameter.

tol::Float64: Tolerance below which convergence is declared.

Examples

julia> reg0001 = createRegressionSetting(@formula(calls ~ year), phones);
julia> ccf(reg0001)
Dict{Any,Any} with 4 entries:
  "betas"     => [-63.4816, 1.30406]
  "outliers"  => [15, 16, 17, 18, 19, 20]
  "lambdas"   => [1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0, 1.0  …  2.77556e-17, 2.77556e-17, 0…
  "residuals" => [-2.67878, -1.67473, -0.37067, -0.266613, 0.337444, 0.941501, 1.44556, 2.04962, 1…

References

Barratt, S., Angeris, G. & Boyd, S. Minimizing a sum of clipped convex functions. Optim Lett 14, 2443–2459 (2020). https://doi.org/10.1007/s11590-020-01565-4

ccf(X, y; starting_lambdas = nothing)

Perform signed gradient descent for clipped convex functions for a given regression setting.

Arguments

X::Array{Float64, 2}: Design matrix of the linear model.

y::Array{Float64, 1}: Response vector of the linear model.

starting_lambdas::Array{Float64,1}: Starting values of weighting parameters used by signed gradient descent.

alpha::Float64: Loss at which a point is labeled as an outlier. If unspecified, will be chosen as p*mean(residuals.^2), where residuals are OLS residuals.

p::Float64: Points that have squared OLS residual greater than p times the mean squared OLS residual are considered outliers.

max_iter::Int64: Maximum number of iterations to run signed gradient descent.

beta::Float64: Step size parameter.

tol::Float64: Tolerance below which convergence is declared.

References

Barratt, S., Angeris, G. & Boyd, S. Minimizing a sum of clipped convex functions. Optim Lett 14, 2443–2459 (2020). https://doi.org/10.1007/s11590-020-01565-4

Atkinson (1994) Forward Search Algorithm

LinRegOutliers.atkinson94 — Function

    atkinson94(setting, iters, crit)

Runs the Atkinson94 algorithm to find out outliers using LMS method.

Arguments

setting::RegressionSetting: A regression setting object.

iters::Int: Number of random samples.

crit::Float64: Critical value for residuals

Examples

julia> reg = createRegressionSetting(@formula(stackloss ~ airflow + watertemp + acidcond), stackloss)
julia> atkinson94(reg)
Dict{Any,Any} with 6 entries:
  "optimum_index"    => 10
  "residuals_matrix" => [0.0286208 0.0620609 … 0.0796249 0.0; 0.0397778 0.120547 … 0.118437 0.0397778; … ; 1.21133 1.80846 … 0.690327 4.14366; 1.61977 0.971592 … 0.616204 3.58098]
  "outliers"         => [1, 3, 4, 21]
  "objective"        => 0.799134
  "coef"             => [-38.3133, 0.745659, 0.432794, 0.0104587]
  "crit"             => 3.0

References

Atkinson, Anthony C. "Fast very robust methods for the detection of multiple outliers." Journal of the American Statistical Association 89.428 (1994): 1329-1339.

## BACON Algorithm (Billor & Hadi & Velleman (2000))

LinRegOutliers.bacon — Function

    bacon(setting, m, method, alpha)

Run the BACON algorithm to detect outliers on regression data.

Arguments:

setting: RegressionSetting object with a formula and a dataset.

m: The number of elements to be included in the initial subset.

method: The distance method to use for selecting the points for initial subset

alpha: The quantile used for cutoff

Examples

julia> reg = createRegressionSetting(@formula(stackloss ~ airflow + watertemp + acidcond), stackloss)
julia> bacon_multivariate_outlier_detection(reg, m=12)
4-element Array{Int64,1}:
  1
  3
  4
 21

References

Billor, Nedret, Ali S. Hadi, and Paul F. Velleman. "BACON: blocked adaptive computationally efficient outlier nominators." Computational statistics & data analysis 34.3 (2000): 279-298.

## Hadi (1994) Algorithm

LinRegOutliers.hadi1994 — Function

hadi1994(multivariateData)

Perform Hadi (1994) algorithm for a given multivariate data.

# Arguments -multivariate::Array{Float64, 2}: Multivariate data.

Examples

julia> multidata = hcat(hbk.x1, hbk.x2, hbk.x3);

julia> hadi1994(multidata)
Dict{Any,Any} with 3 entries:
  "outliers"              => [2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14]
  "critical.chi.squared" => 7.81473
  "rth.robust.distance"   => 5.04541

# Reference Hadi, Ali S. "A modification of a method for the dedection of outliers in multivariate samples" Journal of the Royal Statistical Society: Series B (Methodological) 56.2 (1994): 393-396.

## Chatterjee & Mächler (1997)

LinRegOutliers.cm97 — Function

cm97(setting; maxiter = 1000)

Perform the Chatterjee and Mächler (1997) algorithm for the given regression setting.

Arguments

setting::RegressionSetting: RegressionSetting object with a formula and dataset.

Examples

julia> myreg = createRegressionSetting(@formula(stackloss ~ airflow + watertemp + acidcond), stackloss)
julia> result = cm97(myreg)
Dict{String,Any} with 3 entries:
  "betas"      => [-37.0007, 0.839285, 0.632333, -0.113208]
  "iterations" => 22
  "converged"  => true

References

Chatterjee, Samprit, and Martin Mächler. "Robust regression: A weighted least squares approach." Communications in Statistics-Theory and Methods 26.6 (1997): 1381-1394.